safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.5_The_if_Statement/S12.5_A6_T2-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected keyword 'false'. Expected '(' to start a 'if' condition.
S12.5_A6_T2

PASS Expected parsing failure

TEST COMPLETE
